Add files via upload
This is a code for Air Quality Monitoring in public service vehicles using DHT11 and MQ2 sensors. It involves using an ESP8266 NodeMCU, DHT11, MQ2, Jumper wires and a breadboard. Data is then transferred to a cloud platform: ThingSpeak for visualization. The data collected is Temperature, Humidity and CO2 gas. The code is run on Arduino IDE.